+++ koku/api/report/view.py
@@ -18,23 +18,14 @@
 """View for Reports."""
 import logging
 
-from django.db.models import Value
-from django.db.models.functions import Concat
 from django.utils.translation import ugettext as _
 from pint.errors import DimensionalityError, UndefinedUnitError
 from querystring_parser import parser
 from rest_framework import status
-from rest_framework.decorators import (api_view,
-                                       permission_classes,
-                                       renderer_classes)
-from rest_framework.permissions import AllowAny
 from rest_framework.response import Response
 from rest_framework.serializers import ValidationError
-from rest_framework.settings import api_settings
 
 from api.models import Tenant, User
-from api.report.aws.queries import AWSReportQueryHandler
-from api.report.serializers import QueryParamSerializer
 from api.utils import UnitConverter
 
 LOG = logging.getLogger(__name__)
@@ -210,345 +201,3 @@ def _generic_report(request, provider_parameter_serializer, provider_query_hdlr,
 
     LOG.debug(f'DATA: {output}')
     return Response(output)
-
-
-@api_view(http_method_names=['GET'])
-@permission_classes([AllowAny])
-@renderer_classes(tuple(api_settings.DEFAULT_RENDERER_CLASSES))
-def costs(request):
-    """Get cost data.
-
-    @api {get} /api/v1/reports/costs/ Get cost data
-    @apiName getCostData
-    @apiGroup Report
-    @apiVersion 1.0.0
-    @apiDescription Get cost data.
-
-    @apiHeader {String} token User authorization token.
-    @apiHeader {String} accept HTTP Accept header. (See: RFC2616)
-    @apiHeaderExample {json} Header-Example:
-        {
-            "Accept": "text/csv;q=0.8, application/json"
-        }
-
-    @apiParam (Query Param) {Object} filter The filter to apply to the report.
-    @apiParam (Query Param) {Object} group_by The grouping to apply to the report.
-    @apiParam (Query Param) {Object} order_by The ordering to apply to the report.
-    @apiParamExample {json} Query Param:
-        ?filter[resolution]=daily&filter[time_scope_value]=-10&order_by[cost]=asc
-
-    @apiSuccess {Object} group_by  The grouping to applied to the report.
-    @apiSuccess {Object} order_by  The ordering to applied to the report
-    @apiSuccess {Object} filter  The filter to applied to the report.
-    @apiSuccess {Object} data  The report data.
-    @apiSuccessExample {json} Success-Response:
-        HTTP/1.1 200 OK
-        {
-            "group_by": {
-                "account": [
-                "*"
-                ]
-            },
-            "order_by": {
-                "cost": "asc"
-            },
-            "filter": {
-                "resolution": "daily",
-                "time_scope_value": -10,
-                "time_scope_units": "day",
-                "resource_scope": []
-            },
-            "data": [
-                [
-                {
-                    "date": "2018-05-28",
-                    "accounts": [
-                        {
-                            "account": "8577742690384",
-                            "values": [
-                                {
-                                    "date": "2018-05-28",
-                                    "units": "USD",
-                                    "account": "8577742690384",
-                                    "total": 1498.92962634
-                                }
-                            ]
-                        },
-                        {
-                            "account": "9420673783214",
-                            "values": [
-                                {
-                                    "date": "2018-05-28",
-                                    "units": "USD",
-                                    "account": "9420673783214",
-                                    "total": 1065.845524241
-                                }
-                            ]
-                        }
-                    ]
-                }
-                ]
-            ]
-        }
-    @apiSuccessExample {text} Success-Response:
-        HTTP/1.1 200 OK
-        account,date,total,units
-        6855812392331,2018-07,23008.281583543,USD
-        3028898336671,2018-07,20826.675630200,USD
-        7475489704610,2018-07,20305.483875161,USD
-        2882243055256,2018-07,19474.534357638,USD
-        6721340654404,2018-07,19356.197856632,USD
-
-    """
-    extras = {'report_type': 'costs'}
-    return _generic_report(request, QueryParamSerializer, AWSReportQueryHandler, **extras)
-
-
-@api_view(http_method_names=['GET'])
-@permission_classes([AllowAny])
-@renderer_classes(tuple(api_settings.DEFAULT_RENDERER_CLASSES))
-def instance_type(request):
-    """Get inventory data.
-
-    @api {get} /api/v1/reports/inventory/instance-type/ Get inventory instance type data
-    @apiName getInventoryInstanceTypeData
-    @apiGroup Report
-    @apiVersion 1.0.0
-    @apiDescription Get inventory instance type data.
-
-    @apiHeader {String} token User authorization token.
-    @apiHeader {String} accept HTTP Accept header. (See: RFC2616)
-    @apiHeaderExample {json} Header-Example:
-        {
-            "Accept": "text/csv;q=0.8, application/json"
-        }
-
-    @apiParam (Query Param) {Object} filter The filter to apply to the report.
-    @apiParam (Query Param) {Object} group_by The grouping to apply to the report.
-    @apiParam (Query Param) {Object} order_by The ordering to apply to the report.
-    @apiParam (Query Param) {String} units The units used in the report.
-    @apiParamExample {json} Query Param:
-        ?filter[resolution]=daily&filter[time_scope_value]=-10&order_by[cost]=asc&group_by[account]=*&units=hours
-
-    @apiSuccess {Object} group_by  The grouping to applied to the report.
-    @apiSuccess {Object} filter  The filter to applied to the report.
-    @apiSuccess {Object} data  The data including types, usage, and distinct number of instances for a time period.
-    @apiSuccess {Object} total Aggregates statistics for the report range.
-    @apiSuccessExample {json} Success-Response:
-        HTTP/1.1 200 OK
-        {
-            "group_by": {
-                "account": [
-                "*"
-                ]
-            },
-            "filter": {
-                "resolution": "daily",
-                "time_scope_value": -10,
-                "time_scope_units": "day",
-                "resource_scope": []
-            },
-            "data": [
-                [
-                    {
-                        "date": "2018-05-28",
-                        "accounts": [
-                            {
-                                "account": 111111111111 ,
-                                "instance_types": [
-                                        {
-                                            "instance_type": "t2.medium",
-                                            "values": [
-                                                {
-                                                    "date": "2018-05-28",
-                                                    "units": "Hrs",
-                                                    "instance_type": "t2.medium",
-                                                    "total": 5,
-                                                    "count": 1
-                                                }
-                                            ]
-                                        },
-                                        {
-                                            "instance_type": "m5.2xlarge",
-                                            "values": [
-                                                {
-                                                    "date": "2018-05-28",
-                                                    "units": "Hrs",
-                                                    "instance_type": "m5.2xlarge",
-                                                    "total": 29,
-                                                    "count": 3
-                                                }
-                                            ]
-                                        }
-                                    ]
-                            }
-                        ]
-                    },
-                    {
-                        "date": "2018-05-29",
-                        "accounts": [
-                            {
-                                "account": 111111111111 ,
-                                "instance_types": [
-                                        {
-                                            "instance_type": "m5.2xlarge",
-                                            "values": [
-                                                {
-                                                    "date": "2018-05-28",
-                                                    "units": "Hrs",
-                                                    "instance_type": "m5.2xlarge",
-                                                    "total": 29,
-                                                    "count": 4
-                                                }
-                                            ]
-                                        }
-                                    ]
-                            }
-                        ]
-                    }
-                ]
-            ],
-            "total": {
-                "value": 63,
-                "units": "Hrs",
-                "count": 5
-            }
-        }
-    @apiSuccessExample {text} Success-Response:
-        HTTP/1.1 200 OK
-        account,date,instance_type,total,units
-        3082416796941,2018-08-05,r4.large,11.0,Hrs
-        0840549025238,2018-08-05,m5.large,11.0,Hrs
-        0840549025238,2018-08-05,c5d.2xlarge,9.0,Hrs
-        8133889256380,2018-08-05,c4.xlarge,8.0,Hrs
-        3082416796941,2018-08-04,c5d.2xlarge,12.0,Hrs
-        8133889256380,2018-08-04,c4.xlarge,12.0,Hrs
-        2415722664993,2018-08-04,r4.large,10.0,Hrs
-        8133889256380,2018-08-04,r4.large,10.0,Hrs
-
-    """
-    annotations = {'instance_type':
-                   Concat('cost_entry_product__instance_type', Value(''))}
-    extras = {'annotations': annotations,
-              'group_by': ['instance_type'],
-              'report_type': 'instance_type'}
-    return _generic_report(request, QueryParamSerializer, AWSReportQueryHandler, **extras)
-
-
-@api_view(http_method_names=['GET'])
-@permission_classes([AllowAny])
-@renderer_classes(tuple(api_settings.DEFAULT_RENDERER_CLASSES))
-def storage(request):
-    """Get inventory storage data.
-
-    @api {get} /api/v1/reports/inventory/storage Get inventory storage data
-    @apiName getInventoryStorageData
-    @apiGroup Report
-    @apiVersion 1.0.0
-    @apiDescription Get inventory data.
-
-    @apiHeader {String} token User authorization token.
-
-    @apiParam (Query Param) {Object} filter The filter to apply to the report.
-    @apiParam (Query Param) {Object} group_by The grouping to apply to the report.
-    @apiParam (Query Param) {Object} order_by The ordering to apply to the report.
-    @apiParam (Query Param) {String} units The units used in the report.
-    @apiParamExample {json} Query Param:
-        ?filter[resolution]=daily&filter[time_scope_value]=-10&order_by[cost]=asc&units=byte
-
-    @apiSuccess {Object} group_by  The grouping to applied to the report.
-    @apiSuccess {Object} filter  The filter to applied to the report.
-    @apiSuccess {Object} data  The report data.
-    @apiSuccess {Object} total Aggregates statistics for the report range.
-    @apiSuccessExample {json} Success-Response:
-        HTTP/1.1 200 OK
-        {
-            "group_by": {
-                "account": [
-                    "*"
-                ]
-            },
-            "filter": {
-                "resolution": "monthly",
-                "time_scope_value": "-1",
-                "time_scope_units": "month"
-            },
-            "data": [
-                {
-                    "date": "2018-08",
-                    "accounts": [
-                        {
-                            "account": "0840549025238",
-                            "values": [
-                                {
-                                    "date": "2018-08",
-                                    "units": "GB-Mo",
-                                    "account": "0840549025238",
-                                    "total": 4066.923135971
-                                }
-                            ]
-                        },
-                        {
-                            "account": "3082416796941",
-                            "values": [
-                                {
-                                    "date": "2018-08",
-                                    "units": "GB-Mo",
-                                    "account": "3082416796941",
-                                    "total": 3644.58070225345
-                                }
-                            ]
-                        },
-                        {
-                            "account": "8133889256380",
-                            "values": [
-                                {
-                                    "date": "2018-08",
-                                    "units": "GB-Mo",
-                                    "account": "8133889256380",
-                                    "total": 3584.67567749966
-                                }
-                            ]
-                        },
-                        {
-                            "account": "4783090375826",
-                            "values": [
-                                {
-                                    "date": "2018-08",
-                                    "units": "GB-Mo",
-                                    "account": "4783090375826",
-                                    "total": 3096.66740996526
-                                }
-                            ]
-                        },
-                        {
-                            "account": "2415722664993",
-                            "values": [
-                                {
-                                    "date": "2018-08",
-                                    "units": "GB-Mo",
-                                    "account": "2415722664993",
-                                    "total": 2599.75765963921
-                                }
-                            ]
-                        }
-                    ]
-                }
-            ],
-            "total": {
-                "value": 16992.6045853286,
-                "units": "GB-Mo"
-            }
-        }
-    @apiSuccessExample {text} Success-Response:
-        HTTP/1.1 200 OK
-        account,date,total,units
-        0840549025238,2018-08,4066.923135971,GB-Mo
-        3082416796941,2018-08,3644.58070225345,GB-Mo
-        8133889256380,2018-08,3584.67567749966,GB-Mo
-        4783090375826,2018-08,3096.66740996526,GB-Mo
-        2415722664993,2018-08,2599.75765963921,GB-Mo
-
-    """
-    extras = {'report_type': 'storage'}
-    return _generic_report(request, QueryParamSerializer, AWSReportQueryHandler, **extras)
