+++ koku/api/views.py
@@ -21,4 +21,4 @@
 from api.status.views import status
 from api.iam.view.user_preference import UserPreferenceViewSet
 from api.provider.view import ProviderViewSet
-from api.report.view import costs, instance_type, storage
+from api.report.aws.view import costs, instance_type, storage
